requirements.txt
h5py~=3.4.0
numpy~=1.21.2
PyYAML~=5.4.1
zarr~=2.10.1
netCDF4~=1.5.7